Moved non-interactive flag into submodules.
There are some situations in which a component should error if
it is set to 'non-interactive', for example if a project name is
needed and none has been configured/detected. This patch moves
the non-interactive flag adjudication into the components, so
that they have the freedom to enable this logic.