Added Sphinx-based documentation framework and specs repo.
All of ospenstack's documentation infrastructure is run via sphinx, including
templates, html parsing, uploading, and more. Thus it behooves us to keep
our javascript documentation build as similar as possible. Since we want to
avoid using tox, and littering the project with pythonic artifacts, we're
instead using Sphinx' own recommended build method: A Makefile.

This patch adds a small but complete documentation tree for this project,
including the build tools needed to create it. It satisfies the need for
narrative documentation, but does not (yet) satisfy the need for code
doc generation. It also includes a section for specifications, as this
effort is not yet large enough to warrant its own specification repo.

Code documentation will need to be added in a subsequent patch, using
Sphinx' jsdoc plugins.

Change-Id: I95afbc7001643a3564961c3b5326df4b35c29be8