WIP: Testing nodejs-jobs
This project currently runs a few legacy nodejs jobs.

This change adds new Zuul v3 native nodejs jobs and tests that they
work.

Once the jobs work, we can update the jobs (might be here, might be in
project-config, depending on kind of job).

WIP until jobs are working fine and home of them is clear.

Change-Id: I6a2ca84b7dac396454ac2d88e288382e2ca64920